+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Add a customisation screen to select the lobby/skyblock in Create New World, More Options (warjort)
 * Add an example lobby config (warjort)
 * Save all major config in the persistant data, only skyblocks are re-read after world creation (warjort)
+* Add support for fillers that fill the void with blocks (warjort)
+* Add a stoneblock like example config (warjort)
 
 ### 0.0.1
 * Initial release